+++ content-cache-backends-config/src/state.py
@@ -23,8 +23,10 @@ BACKENDS_CONFIG_NAME = "backends"
 PROTOCOL_CONFIG_NAME = "protocol"
 FAIL_TIMEOUT_CONFIG_NAME = "fail-timeout"
 BACKENDS_PATH_CONFIG_NAME = "backends-path"
-HEALTHCHECK_PATH_CONFIG_NAME = "healthcheck-path"
 HEALTHCHECK_INTERVAL_CONFIG_NAME = "healthcheck-interval"
+HEALTHCHECK_PATH_CONFIG_NAME = "healthcheck-path"
+HEALTHCHECK_SSL_VERIFY_CONFIG_NAME = "healthcheck-ssl-verify"
+HEALTHCHECK_VALID_STATUS_CONFIG_NAME = "healthcheck-valid-status"
 PROXY_CACHE_VALID_CONFIG_NAME = "proxy-cache-valid"
 
 
@@ -97,6 +99,63 @@ def _validate_path_value(value: str) -> str:
     return value
 
 
+class HealthcheckConfig(pydantic.BaseModel):
+    """Represents the configuration for healthchecks.
+
+    Attributes:
+        interval: The time between two healthchecks, in milliseconds.
+        path: The path to check on the backends for health.
+        valid_status: HTTP status codes considered as valid during health checks.
+        ssl_verify: Should we check SSL certificates during health checks.
+    """
+
+    interval: pydantic.PositiveInt
+    path: typing.Annotated[
+        str,
+        pydantic.StringConstraints(min_length=1),
+        pydantic.AfterValidator(_validate_path_value),
+    ]
+    ssl_verify: bool
+    valid_status: tuple[int, ...]
+
+    @classmethod
+    def from_charm(cls, charm: ops.CharmBase) -> "HealthcheckConfig":
+        """Initialize object from the charm.
+
+        Args:
+            charm: The charm containing the configuration.
+
+        Raises:
+            ConfigurationError: Error with the charm configuration.
+
+        Returns:
+            The object.
+        """
+        interval = typing.cast(int, charm.config.get(HEALTHCHECK_INTERVAL_CONFIG_NAME, -1))
+        path = typing.cast(str, charm.config.get(HEALTHCHECK_PATH_CONFIG_NAME, "")).strip()
+        valid_status_str = typing.cast(
+            str, charm.config.get(HEALTHCHECK_VALID_STATUS_CONFIG_NAME, "")
+        ).strip()
+        ssl_verify = typing.cast(bool, charm.config.get(HEALTHCHECK_SSL_VERIFY_CONFIG_NAME, False))
+
+        valid_status = tuple(int(status) for status in valid_status_str.split(","))
+
+        try:
+            # Ignore type check and let pydantic handle the type with validation errors.
+            return cls(
+                interval=interval,
+                path=path,
+                ssl_verify=ssl_verify,  # type: ignore
+                valid_status=valid_status,
+            )
+        except pydantic.ValidationError as err:
+            err_msg = [
+                f'{error["loc"][0]} = {error["input"]}: {error["msg"]}' for error in err.errors()
+            ]
+            logger.error("Found config error: %s", err_msg)
+            raise ConfigurationError(f"Config error: {err_msg}") from err
+
+
 class Configuration(pydantic.BaseModel):
     """Represents the configuration.
 
@@ -107,9 +166,8 @@ class Configuration(pydantic.BaseModel):
         protocol: The protocol to request the backends with. Can be http or https.
         fail_timeout: The time to wait before using a backend after failure.
         backends_path: The path to request the backends.
-        healthcheck_path: The path to request the healthcheck endpoint.
-        healthcheck_interval: The time between two helthchecks, in milliseconds.
         proxy_cache_valid: The cache valid duration.
+        healthcheck: The healthcheck configuration.
     """
 
     hostname: typing.Annotated[
@@ -130,13 +188,8 @@ class Configuration(pydantic.BaseModel):
         pydantic.StringConstraints(min_length=1),
         pydantic.AfterValidator(_validate_path_value),
     ]
-    healthcheck_path: typing.Annotated[
-        str,
-        pydantic.StringConstraints(min_length=1),
-        pydantic.AfterValidator(_validate_path_value),
-    ]
-    healthcheck_interval: pydantic.PositiveInt
     proxy_cache_valid: tuple[str, ...]
+    healthcheck: HealthcheckConfig
 
     @pydantic.field_validator("proxy_cache_valid")
     @classmethod
@@ -183,12 +236,6 @@ class Configuration(pydantic.BaseModel):
             raise ConfigurationError("Empty backends configuration found")
         fail_timeout = typing.cast(str, charm.config.get(FAIL_TIMEOUT_CONFIG_NAME, "")).strip()
         backends_path = typing.cast(str, charm.config.get(BACKENDS_PATH_CONFIG_NAME, "")).strip()
-        healthcheck_path = typing.cast(
-            str, charm.config.get(HEALTHCHECK_PATH_CONFIG_NAME, "")
-        ).strip()
-        healthcheck_interval = typing.cast(
-            int, charm.config.get(HEALTHCHECK_INTERVAL_CONFIG_NAME, -1)
-        )
         proxy_cache_valid_str = typing.cast(
             str, charm.config.get(PROXY_CACHE_VALID_CONFIG_NAME, "")
         ).strip()
@@ -205,6 +252,8 @@ class Configuration(pydantic.BaseModel):
                 f"The proxy_cache_valid is not a list: {proxy_cache_valid_str}"
             )
 
+        healthcheck_config = HealthcheckConfig.from_charm(charm)
+
         try:
             # Ignore type check and let pydantic handle the type with validation errors.
             return cls(
@@ -214,9 +263,8 @@ class Configuration(pydantic.BaseModel):
                 protocol=protocol,  # type: ignore
                 fail_timeout=fail_timeout,
                 backends_path=backends_path,
-                healthcheck_path=healthcheck_path,
-                healthcheck_interval=healthcheck_interval,
                 proxy_cache_valid=proxy_cache_valid,  # type: ignore
+                healthcheck=healthcheck_config,
             )
         except pydantic.ValidationError as err:
             err_msg = [
@@ -245,9 +293,17 @@ class Configuration(pydantic.BaseModel):
                 "Unable to convert configuration to integration data format"
             ) from err
 
+        to_expand = {key for key, value in data.items() if isinstance(value, dict)}
+        for key in to_expand:
+            data.update({f"{key}_{_key}": str(_value) for _key, _value in data[key].items()})
+            data.pop(key)
+
         for key, value in data.items():
             if isinstance(value, str):
+                if value in ["True", "False"]:
+                    data[key] = value.lower()
                 continue
+
             try:
                 data[key] = json.dumps(value)
             except (ValueError, TypeError) as err:
@@ -255,6 +311,7 @@ class Configuration(pydantic.BaseModel):
                 raise ConfigurationError(
                     "Unable to convert configuration to integration data format"
                 ) from err
+
         return data