+++ content-cache-backends-config/tests/unit/test_state.py
@@ -58,8 +58,8 @@ def test_valid_config():
     assert config.protocol == "https"
     assert config.fail_timeout == "30s"
     assert config.backends_path == "/"
-    assert config.healthcheck_path == "/healthz"
-    assert config.healthcheck_interval == 2123
+    assert config.healthcheck.path == "/healthz"
+    assert config.healthcheck.interval == 2123
     assert config.proxy_cache_valid == ()
 
 
@@ -80,8 +80,8 @@ def test_hostname_with_subdomain():
     assert config.protocol == "https"
     assert config.fail_timeout == "30s"
     assert config.backends_path == "/"
-    assert config.healthcheck_path == "/healthz"
-    assert config.healthcheck_interval == 2123
+    assert config.healthcheck.path == "/healthz"
+    assert config.healthcheck.interval == 2123
 
     assert config.proxy_cache_valid == ()
 
@@ -151,7 +151,7 @@ def test_longer_path():
     assert config.protocol == "https"
     assert config.fail_timeout == "30s"
     assert config.backends_path == "/path/to/destination/2"
-    assert config.healthcheck_path == "/path/to/health/2"
+    assert config.healthcheck.path == "/path/to/health/2"
     assert config.proxy_cache_valid == ()
 
 
@@ -253,8 +253,8 @@ def test_http_protocol():
     assert config.protocol == "http"
     assert config.fail_timeout == "30s"
     assert config.backends_path == "/"
-    assert config.healthcheck_path == "/healthz"
-    assert config.healthcheck_interval == 2123
+    assert config.healthcheck.path == "/healthz"
+    assert config.healthcheck.interval == 2123
     assert config.proxy_cache_valid == ()
 
 
@@ -386,8 +386,8 @@ def test_valid_proxy_cache_valid(proxy_cache_valid: str):
     assert config.protocol == "https"
     assert config.fail_timeout == "30s"
     assert config.backends_path == "/"
-    assert config.healthcheck_path == "/healthz"
-    assert config.healthcheck_interval == 2123
+    assert config.healthcheck.path == "/healthz"
+    assert config.healthcheck.interval == 2123
     assert config.proxy_cache_valid == tuple(json.loads(proxy_cache_valid))
 
 
@@ -409,8 +409,10 @@ def test_configuration_to_data():
         "protocol": "https",
         "fail_timeout": "30s",
         "backends_path": "/",
-        "healthcheck_path": "/healthz",
         "healthcheck_interval": "2123",
+        "healthcheck_path": "/healthz",
+        "healthcheck_ssl_verify": "false",
+        "healthcheck_valid_status": "[200]",
         "proxy_cache_valid": "[]",
     }
 
@@ -469,9 +471,7 @@ def test_invalid_healthcheck_path(bad_value, error_msg):
     with pytest.raises(ConfigurationError) as err:
         Configuration.from_charm(charm)
 
-    assert (
-        str(err.value) == f"Config error: ['healthcheck_path = {bad_value.strip()}: {error_msg}']"
-    )
+    assert str(err.value) == f"Config error: ['path = {bad_value.strip()}: {error_msg}']"
 
 
 @pytest.mark.parametrize(
@@ -496,4 +496,4 @@ def test_invalid_healthcheck_interval(bad_value, error_msg):
     with pytest.raises(ConfigurationError) as err:
         Configuration.from_charm(charm)
 
-    assert str(err.value) == f"Config error: ['healthcheck_interval = {bad_value}: {error_msg}']"
+    assert str(err.value) == f"Config error: ['interval = {bad_value}: {error_msg}']"