+++ content-cache/tests/integration/test_healthchecks.py
@@ -17,10 +17,13 @@ from tests.integration.helpers import (
     BACKENDS_PATH_CONFIG_NAME,
     HEALTHCHECK_INTERVAL_CONFIG_NAME,
     HEALTHCHECK_PATH_CONFIG_NAME,
+    HEALTHCHECK_SSL_VERIFY_CONFIG_NAME,
+    HEALTHCHECK_VALID_STATUS_CONFIG_NAME,
     HOSTNAME_CONFIG_NAME,
     PROTOCOL_CONFIG_NAME,
     PROXY_CACHE_VALID_CONFIG_NAME,
     CacheTester,
+    get_app_ip,
 )
 
 HEALTHCHECK_INTERVAL = 2000
@@ -73,6 +76,8 @@ async def test_healthchecks_healthy(
     config[BACKENDS_PATH_CONFIG_NAME] = http_ok_path
     config[HEALTHCHECK_PATH_CONFIG_NAME] = "/health"
     config[HEALTHCHECK_INTERVAL_CONFIG_NAME] = str(HEALTHCHECK_INTERVAL)
+    config[HEALTHCHECK_SSL_VERIFY_CONFIG_NAME] = "false"
+    config[HEALTHCHECK_VALID_STATUS_CONFIG_NAME] = "200"
     config[PROTOCOL_CONFIG_NAME] = "http"
     config[PROXY_CACHE_VALID_CONFIG_NAME] = '["200 10s"]'
     await cache_tester.setup_config(config)
@@ -161,8 +166,6 @@ async def test_healthchecks_one_recovery(
     assert response.status_code == 200
     assert http_ok_message in response.content.decode("utf-8")
 
-    cache_tester._reset_after_run = False
-
 
 @pytest.mark.abort_on_fail
 @pytest.mark.asyncio
@@ -192,3 +195,103 @@ async def test_healthchecks_all_unhealthy(
 
     response = await cache_tester.query_cache(path="/", hostname=hostname, protocol="http")
     assert response.status_code == 502
+
+
+@pytest.mark.parametrize(
+    ["valid_status", "expected_http_code"],
+    [
+        pytest.param("200", 502, id="unhealthy"),
+        pytest.param("200, 418", 200, id="healthy"),
+    ],
+)
+@pytest.mark.abort_on_fail
+@pytest.mark.asyncio
+async def test_healthchecks_custom_status(
+    app: Application,
+    config_app: Application,
+    cache_tester: CacheTester,
+    http_ok_path: str,
+    http_ok_message: str,
+    http_ok_ip: str,
+    valid_status: str,
+    expected_http_code: int,
+    model: Model,
+) -> None:
+    """
+    arrange: One backend responding 418 on its healthcheck. And valid status to match it or not.
+    act: Nothing.
+    assert: HTTP request should fail as 200 is not a valid status here.
+    """
+    hostname = "test.healthchecks.local"
+    config = dict(CacheTester.BASE_CONFIG)
+    config[HOSTNAME_CONFIG_NAME] = hostname
+    config[BACKENDS_CONFIG_NAME] = http_ok_ip
+    config[BACKENDS_PATH_CONFIG_NAME] = http_ok_path
+    config[HEALTHCHECK_PATH_CONFIG_NAME] = "/teapot"
+    config[HEALTHCHECK_INTERVAL_CONFIG_NAME] = str(HEALTHCHECK_INTERVAL)
+    config[HEALTHCHECK_SSL_VERIFY_CONFIG_NAME] = "false"
+    config[HEALTHCHECK_VALID_STATUS_CONFIG_NAME] = valid_status
+    config[PROTOCOL_CONFIG_NAME] = "http"
+    config[PROXY_CACHE_VALID_CONFIG_NAME] = '["200 10s"]'
+    await cache_tester.setup_config(config)
+    await cache_tester.integrate_config()
+    await model.wait_for_idle([app.name, config_app.name], status="active", timeout=10 * 60)
+
+    await asyncio.sleep(5 * HEALTHCHECK_INTERVAL / 1000)
+
+    response = await cache_tester.query_cache(path="/", hostname=hostname, protocol="http")
+    assert response.status_code == expected_http_code
+
+    if expected_http_code == 200:
+        assert http_ok_message in response.content.decode("utf-8")
+
+
+@pytest.mark.parametrize(
+    ["ssl_verify", "expected_http_code"],
+    [
+        pytest.param("false", 200, id="no_ssl_verify"),
+        pytest.param("true", 502, id="ssl_verify"),
+    ],
+)
+@pytest.mark.abort_on_fail
+@pytest.mark.asyncio
+async def test_healthchecks_ssl_verify(
+    app: Application,
+    config_app: Application,
+    cache_tester: CacheTester,
+    http_ok_path: str,
+    http_ok_message: str,
+    https_ok_app: Application,
+    ssl_verify: str,
+    expected_http_code: int,
+    model: Model,
+) -> None:
+    """
+    arrange: One backend responding 418 on its healthcheck. And valid status to match it or not.
+    act: Nothing.
+    assert: HTTP request should fail as 200 is not a valid status here.
+    """
+    https_ok_ip = await get_app_ip(https_ok_app)
+
+    hostname = "test.healthchecks.local"
+    config = dict(CacheTester.BASE_CONFIG)
+    config[HOSTNAME_CONFIG_NAME] = hostname
+    config[BACKENDS_CONFIG_NAME] = https_ok_ip
+    config[BACKENDS_PATH_CONFIG_NAME] = http_ok_path
+    config[HEALTHCHECK_PATH_CONFIG_NAME] = "/health"
+    config[HEALTHCHECK_INTERVAL_CONFIG_NAME] = str(HEALTHCHECK_INTERVAL)
+    config[HEALTHCHECK_SSL_VERIFY_CONFIG_NAME] = ssl_verify
+    config[HEALTHCHECK_VALID_STATUS_CONFIG_NAME] = "200"
+    config[PROTOCOL_CONFIG_NAME] = "https"
+    config[PROXY_CACHE_VALID_CONFIG_NAME] = '["200 10s"]'
+    await cache_tester.setup_config(config)
+    await cache_tester.integrate_config()
+    await model.wait_for_idle([app.name, config_app.name], status="active", timeout=10 * 60)
+
+    await asyncio.sleep(5 * HEALTHCHECK_INTERVAL / 1000)
+
+    response = await cache_tester.query_cache(path="/", hostname=hostname, protocol="http")
+    assert response.status_code == expected_http_code
+
+    if expected_http_code == 200:
+        assert http_ok_message in response.content.decode("utf-8")